+++ src/component/string_component_store.rs
@@ -6,7 +6,7 @@ use core::{
 use std::collections::HashMap;
 
 use crate::error::NotFound;
-use super::{Entity, Component, EntityStore};
+use super::{Entity, Component, EntityStore, ComponentStore};
 
 /// The type key based entity builder is used to create an entity with components.
 pub struct StringEntityBuilder<'a, T> where T: EntityStore
@@ -53,16 +53,18 @@ pub struct StringComponentStore {
     shared: HashMap<Entity, RefCell<HashMap<String, Entity>>>,
 }
 
-impl StringComponentStore {
-    /// Registers an new entity on the store.
-    pub fn register_entity(&mut self, entity: impl Into<Entity>) {
+impl ComponentStore for StringComponentStore {
+    fn register_entity(&mut self, entity: impl Into<Entity>) {
         self.components.insert(entity.into(), HashMap::new());
     }
 
-    /// Removes and entity from the store.
-    pub fn remove_entity(&mut self, entity: impl Into<Entity>) {
+    fn remove_entity(&mut self, entity: impl Into<Entity>) {
         self.components.remove(&entity.into());
     }
+}
+
+impl StringComponentStore {
+   
 
     /// Register a `component` for the given `entity`.
     pub fn register_component<C: Component>(&mut self, entity: Entity, component: C, key: impl Into<String>) {